fix coloring in bash (#94)
ecma48 clears the data buffer when it emits an event. Bash sets multiple styles/colors in the same commend. So, the 2nd+ events emitted for these multi-style sequences had an empty data list, causing them to be ignored by vterm.

IIRC, the original reason for skipping events with empty Raw data was for a bug that was fixed by rewriting the ecma48 parser.